requirements.txt
# Core Dependencies
python>=3.11

# GUI
tkinter

# Voice Recognition & TTS
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.14

# AI & ML
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
openai==1.6.1

# Web & Automation
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
pyautogui==0.9.54

# System Integration
psutil==5.9.6
pywin32==306  # Windows only

# Database
# SQLite is built-in

# Phase 02 Advanced Features
opencv-python==4.8.1.78
mediapipe==0.10.9
face-recognition==1.3.0
screen-brightness-control==0.22.0
pycaw==20231129  # Windows audio control

# Utilities
python-dotenv==1.0.0
